In conda 4.6.11, CONDA_PYTHON_EXE does not exist. Modify to use CONDA_EXE instead.
Rejig the version check to call conda and ask for the version rather than
assuming it based on environment variables.

We previously used CONDA_PYTHON_EXE to determine the location of pythonw.exe
and cwp.py. But it was removed in conda 4.6.11. Now we find pythonw.exe and
cwp.py as relative paths to CONDA_EXE, which points to a known path in
the root environment.

Assumptions are made in either case about where things are located within the
anaconda directory, but this cannot be helped since there is no way to know
where these files and executables are in any canonical way at the moment.