Add subdirectory for art: logos, application icons etc.
Main source is art/source/labscript-suite-icons.svg. It now contains
icons optimised for 16x16, 22x22 and 32x32 resolution, so blurry icons
should be a thing of the past.

This is an inkscape svg with defined grids and rulers and whatnot so
if you're going to edit it, use inkscape.

Banner logos also recreated in this source svg.

Icons and logos are exported at native resolutions and a few multiples
thereof in png, ico and svg form. Any logos you need should be here.

Following Russ's lead, runviewer's icon has red text now.